+++ Sources/App/Middlewares/Security/AIRateLimitMiddleware.swift
@@ -38,7 +38,7 @@ struct AIRateLimitMiddleware: AsyncMiddleware {
     }
     
     func respond(to request: Request, chainingTo next: AsyncResponder) async throws -> Response {
-        let clientIP = extractClientIP(from: request)
+        let clientIP = request.services.ipExtractor.extractClientIP(from: request)
         let currentTime = Date()
         let operationKey = "\(operationType)_\(clientIP)"
         
@@ -95,69 +95,6 @@ struct AIRateLimitMiddleware: AsyncMiddleware {
         response.headers.add(name: "X-RateLimit-Type", value: "AI-\(operationType)")
         
         return response
-    }
-    
-    /// Extracts the real client IP address from the request, checking proxy headers first
-    private func extractClientIP(from request: Request) -> String {
-        // Check X-Forwarded-For header (may contain multiple IPs, client is first)
-        if let forwardedFor = request.headers.first(name: "X-Forwarded-For") {
-            let trimmed = forwardedFor.trimmingCharacters(in: .whitespacesAndNewlines)
-            if !trimmed.isEmpty {
-                // Take the first IP address (the original client)
-                let firstIP = String(trimmed.split(separator: ",").first ?? "")
-                    .trimmingCharacters(in: .whitespacesAndNewlines)
-                if isValidIPAddress(firstIP) {
-                    return firstIP
-                }
-            }
-        }
-        
-        // Check X-Real-IP header (single IP)
-        if let realIP = request.headers.first(name: "X-Real-IP") {
-            let trimmed = realIP.trimmingCharacters(in: .whitespacesAndNewlines)
-            if !trimmed.isEmpty && isValidIPAddress(trimmed) {
-                return trimmed
-            }
-        }
-        
-        // Check CF-Connecting-IP header (Cloudflare specific)
-        if let cfIP = request.headers.first(name: "CF-Connecting-IP") {
-            let trimmed = cfIP.trimmingCharacters(in: .whitespacesAndNewlines)
-            if !trimmed.isEmpty && isValidIPAddress(trimmed) {
-                return trimmed
-            }
-        }
-        
-        // Fallback to remote address
-        return request.remoteAddress?.hostname ?? "unknown"
-    }
-    
-    /// Validates if a string is a valid IP address (IPv4 or IPv6)
-    private func isValidIPAddress(_ ip: String) -> Bool {
-        guard !ip.isEmpty, ip != "unknown", ip != "localhost" else { return false }
-        
-        // Check for IPv4 format
-        if ip.contains(".") {
-            let components = ip.split(separator: ".")
-            guard components.count == 4 else { return false }
-            return components.allSatisfy { component in
-                guard let num = Int(component), num >= 0, num <= 255 else { return false }
-                return true
-            }
-        }
-        
-        // Basic IPv6 validation
-        if ip.contains(":") {
-            let components = ip.split(separator: ":")
-            guard components.count >= 2, components.count <= 8 else { return false }
-            return components.allSatisfy { component in
-                return component.allSatisfy { char in
-                    char.isHexDigit
-                }
-            }
-        }
-        
-        return false
     }
 }
