+++ Sources/App/Modules/CacheAdmin/CacheAdminModule.swift
@@ -1,0 +1,9 @@
+import Vapor
+
+struct CacheAdminModule: ModuleInterface {
+    let router = CacheAdminRouter()
+    
+    func boot(_ app: Application) throws {
+        try router.boot(routes: app.routes)
+    }
+}
