+++ Sources/App/Modules/RulesGeneration/RulesGenerationRouter.swift
@@ -2,7 +2,6 @@ import Vapor
 
 struct RulesGenerationRouter: RouteCollection {
     let controller = RulesGenerationController()
-    let cacheAdminController = AICacheAdminController()
     
     func boot(routes: any RoutesBuilder) throws {
         let api = routes
@@ -20,21 +19,5 @@ struct RulesGenerationRouter: RouteCollection {
         // Rules generation endpoint with moderate rate limiting (10/hour)
         let rulesGenerationAPI = api.grouped(AIRateLimitMiddleware(operationType: .rulesGeneration))
         rulesGenerationAPI.post("rules-summary", use: controller.generateRulesSummary)
-        
-        // Admin cache management endpoints (requires admin authentication)
-        let adminAPI = routes
-            .grouped("api")
-            .grouped("admin")
-            .grouped("cache")
-            .grouped(EnsureAdminUserMiddleware())
-        
-        // Cache statistics and monitoring
-        adminAPI.get("stats", use: cacheAdminController.getCacheStatistics)
-        adminAPI.get("health", use: cacheAdminController.getCacheHealth)
-        adminAPI.get("entries", use: cacheAdminController.getCacheEntries)
-        
-        // Cache management operations
-        adminAPI.delete(use: cacheAdminController.clearCache)
-        adminAPI.post("cleanup", use: cacheAdminController.manualCleanup)
     }
 }
